Do not actually check BE0 on writes after the first word
Since we are flowing into the next word, the low byte will _always_ be
true. We do need to check the high byte in case it is the last byte of
the transaction.

Eventually, I will update the code to only check BE1 if it is burst
last. Not now though