Compact code down into read8 and write8 routines
We want to make this code as simple as possible. Use the timing
information gleaned from the CH351 datasheet via copilot. Turns out I
have been playing with fire for a very long time with respect to the
CH351s. What is very interesting is that the extra hold times explains
why I had to introduce delay states for the 2560 EBI.