Was able to remove 200ns breathe time from getAddress
The 50ns of breathe is already taken into account in read8. This doesn't
provide a massive speed boost but a slight improvement.

Also, added a parameter for readDataLines to be more selective when
reading from the CH351. If careful, it can eliminate 250ns in some cases
at the slight cost of evaluation.